Move `branding` to features
