+++ apps/console/src/features/branding/__tests__/__mocks__/branding-permissions.ts
