+++ apps/console/src/features/branding/components/design/__tests__/theme-swatch.test.tsx
