+++ apps/console/src/features/branding/components/design/illustrations-preview.tsx
