+++ apps/console/src/features/branding/components/design/layout-design-extensions/layout-design-extension-component-interface.tsx
