+++ apps/console/src/features/branding/components/design/layout-design-extensions/side-aligned.tsx
