+++ apps/console/src/features/branding/components/design/layout-swatch/index.ts
