+++ apps/console/src/features/branding/components/design/layout-swatch/layout-swatch-adapter.tsx
