+++ apps/console/src/features/branding/components/design/theme-swatch/theme-swatch-adapter.tsx
