+++ apps/console/src/features/branding/components/general/general-details-form.tsx
@@ -22,7 +22,7 @@ import React, { FunctionComponent, MutableRefObject, ReactElement, Ref, forwardR
 import { useTranslation } from "react-i18next";
 import { useSelector } from "react-redux";
 import { Placeholder } from "semantic-ui-react";
-import { AppState } from "../../../../../features/core/store";
+import { AppState } from "../../../core/store";
 import { BrandingPreferencesConstants } from "../../constants";
 import { BrandingPreferenceInterface } from "../../models";
