+++ apps/console/src/features/branding/components/preview/product-header.tsx
