+++ apps/console/src/features/branding/components/sticky-tab-pane-action-panel.tsx
