+++ apps/console/src/features/branding/constants/branding-preferences-constants.ts
