+++ apps/console/src/features/branding/constants/index.ts
