+++ apps/console/src/features/branding/meta/layouts/side-aligned.ts
