+++ apps/console/src/features/branding/meta/layouts/side-image.ts
