+++ apps/console/src/features/branding/meta/themes/dark-theme.ts
