+++ apps/console/src/features/branding/meta/themes/light-theme.ts
