+++ apps/console/src/features/branding/utils/index.ts
